수정

+++ Furnifit/src/main/java/com/furnifit/member/domain/Coupon.java
@@ -2,7 +2,7 @@ package com.furnifit.member.domain;
 
 public class Coupon {
 	private int serial;
-	private double discountRate;
+	private int discountRate;
 	private String expirationDate;
 	private String status;
 	
@@ -12,10 +12,10 @@ public class Coupon {
 	public void setSerial(int serial) {
 		this.serial = serial;
 	}
-	public double getDiscountRate() {
+	public int getDiscountRate() {
 		return discountRate;
 	}
-	public void setDiscountRate(double discountRate) {
+	public void setDiscountRate(int discountRate) {
 		this.discountRate = discountRate;
 	}
 	public String getExpirationDate() {
